adding custom exception and response

+++ HR.Application/Features/LeaveRequests/Handlers/Commands/DeleteLeaveRequestCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
+using HR.Application.Exceptions;
 using HR.Application.Features.LeaveRequests.Requests.Commands;
 using HR.Application.Persistance.Contracts;
+using HR.Domain;
 using MediatR;
 
 namespace HR.Application.Features.LeaveRequests.Handlers.Commands
@@ -18,6 +20,10 @@ namespace HR.Application.Features.LeaveRequests.Handlers.Commands
         public async Task<Unit> Handle(DeleteLeaveRequestCommand request, CancellationToken cancellationToken)
         {
             var leaveRequest = await _leaveRequestRepository.Get(request.Id);
+            if (leaveRequest == null)
+            {
+                throw new NotFoundEception(nameof(LeaveRequest), request.Id);
+            }
             await _leaveRequestRepository.Delete(leaveRequest);
             return Unit.Value;
         }
