+++ HR.Application/Features/LeaveTypes/Handlers/Commands/DeleteLeaveTypeCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
+using HR.Application.Exceptions;
 using HR.Application.Features.LeaveTypes.Requests.Commands;
 using HR.Application.Persistance.Contracts;
+using HR.Domain;
 using MediatR;
 
 namespace HR.Application.Features.LeaveTypes.Handlers.Commands
@@ -19,6 +21,10 @@ namespace HR.Application.Features.LeaveTypes.Handlers.Commands
         public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
         {
             var leaveType = await _leaveTypeRepository.Get(request.Id);
+            if (leaveType == null)
+            {
+                throw new NotFoundEception(nameof(LeaveType), request.Id);
+            }
             await _leaveTypeRepository.Delete(leaveType);
             return Unit.Value;
         }
